fixed some complier warnings

+++ Sender/OpenerViewController.m
@@ -75,21 +75,21 @@
     autoPush = YES;
     tickCount = startRecordTime;
     [self.timeLabel setTextColor:[UIColor whiteColor]];
-    [self.timeLabel setText:[NSString stringWithFormat:@"Reply in 0:%02d",tickCount]];
+    [self.timeLabel setText:[NSString stringWithFormat:@"Reply in 0:%02ld",(long)tickCount]];
     self.startRecordTimer = [NSTimer scheduledTimerWithTimeInterval:1 target:self selector:@selector(startRecordingWithTimer:) userInfo:nil repeats:YES];
  
 }
 
 - (void)startRecordingWithTimer:(NSTimer*)timer
 {
-    [self.timeLabel setText:[NSString stringWithFormat:@"Reply in 0:%02d",tickCount]];
+    [self.timeLabel setText:[NSString stringWithFormat:@"Reply in 0:%02ld",(long)tickCount]];
     [self.timeLabel setTextColor:[UIColor whiteColor]];
     tickCount--;
     if(tickCount <= 0)
     {
         tickCount = self.videoPlayerVC.videoLength - startRecordTime;
         [self.timeLabel setTextColor:[UIColor redColor]];
-        [self.timeLabel setText:[NSString stringWithFormat:@"Recording Reaction 0:%02d",tickCount]];
+        [self.timeLabel setText:[NSString stringWithFormat:@"Recording Reaction 0:%02ld",(long)tickCount]];
         self.reactionTimer = [NSTimer scheduledTimerWithTimeInterval:1 target:self selector:@selector(onReactionTick:) userInfo:nil repeats:YES];
         NSAssert([timer isEqual:self.startRecordTimer], @"expecting timer to equal startRecordingTimer");
         [self.startRecordTimer invalidate];
@@ -183,7 +183,7 @@
 {
     tickCount = MAX_RECORD_TIME;
     [self.timeLabel setTextColor:[UIColor redColor]];
-    [self.timeLabel setText:[NSString stringWithFormat:@"Recording 0:%02d",tickCount]];
+    [self.timeLabel setText:[NSString stringWithFormat:@"Recording 0:%02ld",(long)tickCount]];
     
     if (startTimer) {
         [self.reactionTimer invalidate];
@@ -196,7 +196,7 @@
 {
     tickCount--;
     [self.timeLabel setTextColor:[UIColor redColor]];
-    [self.timeLabel setText:[NSString stringWithFormat:@"Recording Reaction 0:%02d",tickCount]];
+    [self.timeLabel setText:[NSString stringWithFormat:@"Recording Reaction 0:%02ld",(long)tickCount]];
 }
 
 - (void)onTick:(NSTimer*)timer
@@ -207,7 +207,7 @@
         tickCount = 0;
     }
     [self.timeLabel setTextColor:[UIColor redColor]];
-    [self.timeLabel setText:[NSString stringWithFormat:@"Recording 0:%02d",tickCount]];
+    [self.timeLabel setText:[NSString stringWithFormat:@"Recording 0:%02ld",(long)tickCount]];
     
 }
